refactor(ppl-lint): cut command-suggestion + dead quick-fix path
Shrinks the PR for reviewability without losing either shipped rule
(head-without-sort, division-by-zero):

- Remove the "did you mean" command-typo suggestion (command_suggestion
  + its two drift-alarm guard tests).
- Remove the quick-fix subsystem (fix_registry, code_action_provider,
  the syntax/lint fix side-tables in language.ts, DiagnosticFix). Neither
  shipped rule emits a fix, so getModelFix was always undefined.
- Relocate markerFixKey (a position+message key-builder, unrelated to
  fixes) into hover_registry.ts so hover keeps working after fix_registry
  is deleted.
- Drop inert tests that import no shipped source (field_slot_grammar_guard,
  doc_links + fixture).
- Reset routes/index.ts to the ppl-lint-backend canonical form (the
  definePPLBundleRoute refactor was duplicate churn).

Hover tooltips and lint squiggles for both rules are retained.
