+++ packages/osd-monaco/src/ppl/language.test.ts
@@ -29,7 +29,6 @@ jest.mock('../monaco', () => ({
       setMonarchTokensProvider: jest.fn(),
       setTokensProvider: jest.fn(),
       registerCompletionItemProvider: jest.fn(),
-      registerCodeActionProvider: jest.fn(),
       registerHoverProvider: jest.fn(),
       registerDocumentRangeFormattingEditProvider: jest.fn(),
     },
@@ -72,12 +71,8 @@ jest.mock('./validation_provider', () => ({
 jest.mock('./lint/diagnostic_to_marker', () => ({
   diagnosticToMarker: (d: { ruleId: string }) => ({ message: d.ruleId, code: d.ruleId }),
 }));
-jest.mock('./lint/fix_registry', () => ({
-  markerFixKey: (m: { code: string }) => m.code,
-  setModelFixes: jest.fn(),
-  clearModelFixes: jest.fn(),
-}));
 jest.mock('./lint/hover/hover_registry', () => ({
+  markerFixKey: (m: { code: string }) => m.code,
   setModelHoverFacts: jest.fn(),
   clearModelHoverFacts: jest.fn(),
 }));
