+++ packages/osd-monaco/src/ppl/lint/diagnostic_to_marker.ts
@@ -27,8 +27,6 @@ function toMonacoRange(range: DiagnosticRange): MonacoRange {
 
 export const LINT_MARKER_SOURCE = 'ppl-lint';
 
-export const SYNTAX_MARKER_SOURCE = 'ppl-syntax';
-
 function toMarkerSeverity(severity: LintSeverity): monaco.MarkerSeverity {
   switch (severity) {
     case 'error':
@@ -60,16 +58,6 @@ export function diagnosticToMarker(diagnostic: Diagnostic): monaco.editor.IMarke
     marker.code = diagnostic.docUrl
       ? { value: diagnostic.ruleId, target: monaco.Uri.parse(diagnostic.docUrl) }
       : diagnostic.ruleId;
-  }
-
-  if (diagnostic.fix) {
-    (marker as monaco.editor.IMarkerData & {
-      fix?: { title: string; text: string; range?: MonacoRange };
-    }).fix = {
-      title: diagnostic.fix.title,
-      text: diagnostic.fix.text,
-      range: diagnostic.fix.range ? toMonacoRange(diagnostic.fix.range) : undefined,
-    };
   }
 
   if (diagnostic.hoverFacts) {
