+++ packages/osd-monaco/src/ppl/ppl_error_listener.ts
@@ -10,7 +10,6 @@ import {
   Token,
   ATNSimulator,
 } from 'antlr4ng';
-import { buildCommandSuggestion, CommandSuggestion } from './command_suggestion';
 
 export interface SyntaxError {
   message: string;
@@ -18,18 +17,6 @@ export interface SyntaxError {
   column: number;
   endLine?: number;
   endColumn?: number;
-  /**
-   * Stable machine-readable identity for a recognized diagnostic. Absent for
-   * raw ANTLR errors. Reuses {@link CommandSuggestion}'s `code` so the two stay
-   * in lock-step if a second code appears.
-   */
-  code?: CommandSuggestion['code'];
-  /**
-   * Structured, deterministic correction that drives a Monaco quick-fix
-   * lightbulb. Mirrors {@link CommandSuggestion}'s `fix` (and the lint path's
-   * `Diagnostic.fix`). Absent when there is no unambiguous rewrite.
-   */
-  fix?: CommandSuggestion['fix'];
 }
 
 export class PPLSyntaxErrorListener implements ANTLRErrorListener {
@@ -49,15 +36,8 @@ export class PPLSyntaxErrorListener implements ANTLRErrorListener {
       endColumn = charPositionInLine + offendingSymbol.text.length;
     }
 
-    // When the offending token is a misspelled command, replace ANTLR's noisy
-    // "mismatched input ... expecting {40 keywords}" message with a friendly
-    // suggestion and attach a one-click fix. Falls back to the raw message.
-    const suggestion = buildCommandSuggestion(recognizer, offendingSymbol, e);
-
     this.errors.push({
-      message: suggestion?.message ?? msg,
-      code: suggestion?.code,
-      fix: suggestion?.fix,
+      message: msg,
       line,
       column: charPositionInLine,
       endLine: line,
