+++ src/plugins/query_enhancements/server/routes/index.ts
@@ -142,6 +142,10 @@ export function defineSearchStrategyRouteProvider(logger: Logger, router: IRoute
   };
 }
 
+/**
+ * Defines route for PPL grammar bundle endpoint
+ * Forwards requests to OpenSearch /_plugins/_ppl/_grammar
+ */
 export function definePPLBundleRoute(logger: Logger, router: IRouter) {
   router.get(
     {
@@ -155,10 +159,15 @@ export function definePPLBundleRoute(logger: Logger, router: IRouter) {
     async (context, req, res): Promise<IOpenSearchDashboardsResponse<any | ResponseError>> => {
       try {
         const { dataSourceId } = req.query;
-        const opensearchClient = await resolveOpenSearchClient(context, dataSourceId);
-        if (!opensearchClient) {
-          return res.custom({ statusCode: 400, body: DATASOURCE_UNAVAILABLE_MESSAGE });
+        if (dataSourceId && !context.dataSource?.opensearch?.getClient) {
+          return res.custom({
+            statusCode: 400,
+            body: 'dataSourceId is not supported because data source plugin is unavailable',
+          });
         }
+        const opensearchClient = dataSourceId
+          ? await context.dataSource.opensearch.getClient(dataSourceId)
+          : context.core.opensearch.client.asCurrentUser;
         const result = await opensearchClient.transport.request({
           method: 'GET',
           path: URI.PPL_BUNDLE,
